Rename the agentSession directory to agent-session.
Deep-import paths will mirror the published file tree once dist/ becomes
the package root, so the directory name must literally match the public
subpath ./agent-session (legacy moduleResolution cannot follow aliases).
