+++ packages/harness/src/agent-session/builtinsFromSpec.ts
