+++ packages/harness/src/agent-session/schemas/agentSpec.ts
