+++ packages/harness/src/agent-session/store/ISessionStore.ts
