+++ packages/harness/src/index.ts
@@ -2,5 +2,5 @@
  * Root barrel for @truefoundry/utils.
  * Namespaced to avoid colliding event-schema exports between layers.
  */
-export * as agentSession from './agentSession';
+export * as agentSession from './agent-session';
 export * as core from './core';
